Update README.md
This Android project is a feature-rich Music Player App developed in Kotlin using Android Studio.
It not only plays music stored in the device but also supports audio recording using the device’s microphone.

The user can play, pause, and switch to the next song, and also record custom audio which can be stored and later accessed.

The app uses Android’s built-in MediaPlayer for playback and MediaRecorder for audio recording. It also handles runtime permissions for microphone and storage access.